mv lora test
